+++ includes/filetree/filetree.php
@@ -2,7 +2,7 @@
 /**
 *
 * @package Upload Extensions
-* @copyright (c) 2014 - 2017 Igor Lavrov (https://github.com/LavIgor) and John Peskens (http://ForumHulp.com)
+* @copyright (c) 2014 - 2019 Igor Lavrov (https://github.com/LavIgor) and John Peskens (http://ForumHulp.com)
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -82,7 +82,6 @@ class filetree
 			if ($first_call)
 			{
 				$php_file_tree .= ' class="php-file-tree"';
-				$first_call = false;
 			}
 			$php_file_tree .= '>';
 			foreach ($file as $this_file)