+++ includes/functions/extensions.php
@@ -2,7 +2,7 @@
 /**
 *
 * @package Upload Extensions
-* @copyright (c) 2014 - 2017 Igor Lavrov (https://github.com/LavIgor) and John Peskens (http://ForumHulp.com)
+* @copyright (c) 2014 - 2019 Igor Lavrov (https://github.com/LavIgor) and John Peskens (http://ForumHulp.com)
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -15,47 +15,7 @@ use \boardtools\upload\includes\filetree\filedownload;
 class extensions
 {
 	/**
-	* Check the version and return the available updates.
-	*
-	* @param \phpbb\extension\metadata_manager $md_manager The metadata manager for the version to check.
-	* @param bool $force_update Ignores cached data. Defaults to false.
-	* @param bool $force_cache Force the use of the cache. Override $force_update.
-	* @return string
-	* @throws RuntimeException
-	*/
-	public static function version_check(\phpbb\extension\metadata_manager $md_manager, $force_update = false, $force_cache = false)
-	{
-		$cache = objects::$cache;
-		$config = objects::$config;
-		$user = objects::$user;
-		$meta = $md_manager->get_metadata('all');
-
-		if (!isset($meta['extra']['version-check']))
-		{
-			throw new \RuntimeException($user->lang('NO_VERSIONCHECK'), 1);
-		}
-
-		$version_check = $meta['extra']['version-check'];
-
-		if (version_compare($config['version'], '3.1.1', '>'))
-		{
-			$version_helper = new \phpbb\version_helper($cache, $config, new \phpbb\file_downloader(), $user);
-		}
-		else
-		{
-			$version_helper = new \phpbb\version_helper($cache, $config, $user);
-		}
-		$version_helper->set_current_version($meta['version']);
-		$version_helper->set_file_location($version_check['host'], $version_check['directory'], $version_check['filename']);
-		$version_helper->force_stability($config['extension_force_unstable'] ? 'unstable' : null);
-
-		return $updates = $version_helper->get_suggested_updates($force_update, $force_cache);
-	}
-
-	/**
 	* Lists all the available extensions and dumps to the template
-	*
-	* @return null
 	*/
 	public static function list_uninstalled_exts()
 	{
@@ -112,8 +72,6 @@ class extensions
 
 	/**
 	* Lists all the extensions and dumps to the template
-	*
-	* @return null
 	*/
 	public static function list_all_exts()
 	{
@@ -136,7 +94,7 @@ class extensions
 				);
 
 				$force_update = objects::$request->variable('versioncheck_force', false);
-				$updates = self::version_check($md_manager, $force_update, !$force_update);
+				$updates = objects::$compatibility->version_check($md_manager, $force_update, !$force_update, objects::$config['extension_force_unstable'] ? 'unstable' : null);
 
 				$extension_meta_data[$name]['S_UP_TO_DATE'] = empty($updates);
 				$extension_meta_data[$name]['S_VERSIONCHECK'] = true;
@@ -189,6 +147,9 @@ class extensions
 
 	/**
 	* Sort helper for the table containing the metadata about the extensions.
+	* @param array $val1 First metadata array
+	* @param array $val2 Second metadata array
+	* @return int
 	*/
 	protected static function sort_extension_meta_data_table($val1, $val2)
 	{
@@ -509,7 +470,6 @@ class extensions
 	/**
 	* Checks availability of updates for the specified extension.
 	* @param string $ext_name The name of the extension.
-	* @return null
 	*/
 	public static function ajax_versioncheck($ext_name)
 	{
@@ -517,15 +477,16 @@ class extensions
 
 		try
 		{
-			$meta = $md_manager->get_metadata('all');
+			// Validate extension's metadata
+			$md_manager->get_metadata('all');
 
 			$force_update = true;
-			$updates = self::version_check($md_manager, $force_update, !$force_update);
+			$updates_available = objects::$compatibility->version_check($md_manager, $force_update, !$force_update, objects::$config['extension_force_unstable'] ? 'unstable' : null);
 
 			self::response(array(
 				'ext_name'		=> $ext_name,
 				'status'		=> 'success',
-				'versioncheck'	=> (empty($updates)) ? "up_to_date" : "not_up_to_date",
+				'versioncheck'	=> (empty($updates_available)) ? "up_to_date" : "not_up_to_date",
 				'message'		=> objects::$user->lang(empty($updates_available) ? 'UP_TO_DATE' : 'NOT_UP_TO_DATE', $md_manager->get_metadata('display-name'))
 			));
 		}
@@ -553,7 +514,7 @@ class extensions
 	/**
 	* Creates a ZIP package of the extension and prepares it for downloading.
 	* @param string $ext_name The name of the extension.
-	* @return null|bool
+	* @return bool
 	*/
 	public static function download_extension($ext_name)
 	{
@@ -622,13 +583,15 @@ class extensions
 				fclose($fp);
 			}
 		}
+
+		return true;
 	}
 
 	/**
 	* Gets missing language directories for an extension from a specified zip file.
 	* @param string $ext_name The name of the extension.
 	* @param string $zip_file The name of zip file.
-	* @return null|bool
+	* @return bool
 	*/
 	public static function restore_languages($ext_name, $zip_file)
 	{
@@ -675,5 +638,7 @@ class extensions
 		}
 
 		files::catch_errors(files::rrmdir($ext_tmp));
+
+		return true;
 	}
 }